[FEATURE] Allow to re-import data marked as "NEW" records
A new option in a l10n configuration allows for a new behaviour in the following use-cases

a) If a new translation for records are created and external translation is in progress, however
some elements are translated in the meantime within TYPO3, but the translations (which could have
a higher priority) cannot override the existing datasets at all anymore, because the NEW parameter
within the XML does not use the DataMap but goes with the CommandMap in DataHandler, thus
having no effect. The option effectively maps the translated data of the XML into the DataMap part
to update existing records then if they have been translated.

b) If an XML file was exported with new records to translate, and re-imported again in the system,
but then there is a v2 of the XML file, the import won't have any effect, because the "NEW" parameter
is skipped for the reasons above (translation already exists). With the new option activated the
data is always updated then when importing the same file multiple times, even if the translations
were marked as "NEW".

Resolves: #77272
Releases: master, 5.0
Change-Id: Icaa56acb85edc453c6afd1fc151141954989cb41
Reviewed-on: https://review.typo3.org/49230
Reviewed-by: Jo Hasenau <info@cybercraft.de>
Tested-by: Jo Hasenau <info@cybercraft.de>